refactor(P6spy): 정책에 따른 P6Spy 관련 설정 수정
