+++ src/main/resources/application-test.yml
@@ -20,7 +20,7 @@ spring:
       ddl-auto: none
     properties:
       hibernate:
-        dialect: org.hibernate.spatial.dialect.mysql.MySQL8SpatialDialect
+        dialect: org.hibernate.spatial.dialect.mysql.MySQLSpatialDialect
 
   flyway:
     enabled: false
@@ -44,8 +44,3 @@ decorator:
   datasource:
     p6spy:
       enable-logging: true
-
-logging:
-  level:
-    org.hibernate.SQL: debug
-    org.hibernate.type.descriptor.sql: trace
