+++ src/test/spy.properties
@@ -1,0 +1,1 @@
+logMessageFormat=com.prgrms.mukvengers.global.config.p6spy.P6spyMessageFormatter
